# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TaskStream API",
        "description": "Dokumentasi RESTful API untuk TaskStream - Platform Manajemen Tugas dengan Fitur Real-time.",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development API Server"
        }
    ],
    "paths": {
        "/tasks/{taskId}/attachments": {
            "post": {
                "tags": [
                    "Attachments"
                ],
                "summary": "Upload Attachment",
                "description": "Unggah file lampiran untuk tugas. Mendukung standard upload (file <= 50MB) dan chunked upload (file > 50MB).",
                "operationId": "00c7eb9124b818bc358bc61214ba8ef0",
                "responses": {
                    "201": {
                        "description": "File berhasil diunggah"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/attachments/{id}/download": {
            "get": {
                "tags": [
                    "Attachments"
                ],
                "summary": "Download / Stream Attachment",
                "description": "Unduh berkas lampiran secara langsung. Jika berkas berupa video (e.g. MP4/WebM), mendukung pemutaran range-requests (206 Partial Content) untuk seek player.",
                "operationId": "999ae10d9bba8983e768670cbe6adc86",
                "responses": {
                    "200": {
                        "description": "Download berkas / Ranged stream partial content"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/attachments/{id}": {
            "delete": {
                "tags": [
                    "Attachments"
                ],
                "summary": "Delete Attachment",
                "description": "Hapus berkas lampiran beserta seluruh riwayat versi fisiknya.",
                "operationId": "f4c4544cd9697f064ae62eaeeacafbf9",
                "responses": {
                    "200": {
                        "description": "Attachment berhasil dihapus"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Gunakan bearer access_token yang didapatkan dari login.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}